// src/src/app/Models/bwUsers.ts
export class BwUsers {
    _id: string = '';
    moNumber: string = '';
    adharNumber: string = '';
    firstName: string = '';
    middleName: string = '';
    lastName: string = '';
    fatherName: string = '';
    address1: string = '';
    address2: string = '';
    address3: string = '';
    area: string = null;
    city: string = '';
    date_of_birth: string = '';
    state: string = '';
    pincode: string = '';
    SKILLS: string = '';
    CORE_SKILLS: any[] = [];
    postal_area: any = '';
    WORK_EXPERIENCE: any[] = [];
    imagePath: any[] = [];
}